# server/internal/transport/http/handlers/auth.go
package handlers

import (
	"net/http"
	"strings"

	"github.com/gin-gonic/gin"

	"github.com/shenfay/kiqi/internal/app/authentication"
	"github.com/shenfay/kiqi/internal/transport/http/response"
	"github.com/shenfay/kiqi/pkg/errors"
	userErr "github.com/shenfay/kiqi/pkg/errors/user"
	validationErr "github.com/shenfay/kiqi/pkg/errors/validation"
)

// AuthHandler 认证 HTTP 处理器
type AuthHandler struct {
	service      *authentication.Service
	tokenService authentication.TokenService
}

// NewAuthHandler 创建认证处理器实例
func NewAuthHandler(service *authentication.Service, tokenService authentication.TokenService) *AuthHandler {
	return &AuthHandler{
		service:      service,
		tokenService: tokenService,
	}
}

// RegisterRequest 用户注册请求
type RegisterRequest struct {
	Email    string `json:"email" binding:"required,email,max=255"`   // 用户邮箱
	Password string `json:"password" binding:"required,min=8,max=72"` // 用户密码（8-72字符）
}

// LoginRequest 用户登录请求
type LoginRequest struct {
	Email    string `json:"email" binding:"required,email"` // 用户邮箱
	Password string `json:"password" binding:"required"`    // 用户密码
}

// RefreshTokenRequest 刷新 Token 请求
type RefreshTokenRequest struct {
	RefreshToken string `json:"refresh_token" binding:"required"` // 刷新令牌
}

// Register 处理用户注册
//
// 创建新用户账户并返回认证令牌。
// 邮箱必须在系统中唯一。
//
// @Summary Register a new user account
// @Tags Authentication
// @Accept json
// @Produce json
// @Param request body RegisterRequest true "用户注册数据"
// @Success 201 {object} response.SuccessResponse{data=authentication.AuthResponse} "注册成功"
// @Failure 400 {object} response.ErrorResponse "请求参数错误"
// @Failure 409 {object} response.ErrorResponse "邮箱已存在"
// @Failure 500 {object} response.ErrorResponse "服务器内部错误"
// @Router /auth/register [post]
func (h *AuthHandler) Register(c *gin.Context) {
	var req RegisterRequest
	if err := c.ShouldBindJSON(&req); err != nil {
		response.Error(c, validationErr.FromGinError(err))
		return
	}

	cmd := authentication.RegisterCommand{
		Email:    req.Email,
		Password: req.Password,
	}

	resp, err := h.service.Register(c.Request.Context(), cmd)
	if err != nil {
		response.Error(c, err)
		return
	}

	response.Created(c, authentication.ToAuthResponse(resp))
}

// Login 处理用户登录
//
// 验证用户凭据并返回访问/刷新令牌。
// 跟踪登录失败次数，失败过多时锁定账户。
//
// @Summary 用户登录并返回令牌
// @Tags Authentication
// @Accept json
// @Produce json
// @Param request body LoginRequest true "Login credentials"
// @Success 200 {object} response.SuccessResponse{data=authentication.AuthResponse} "登录成功"
// @Failure 400 {object} response.ErrorResponse "请求参数错误"
// @Failure 401 {object} response.ErrorResponse "账号或密码错误"
// @Failure 423 {object} response.ErrorResponse "账户已锁定"
// @Failure 500 {object} response.ErrorResponse "服务器内部错误"
// @Router /auth/login [post]
func (h *AuthHandler) Login(c *gin.Context) {
	var req LoginRequest
	if err := c.ShouldBindJSON(&req); err != nil {
		response.Error(c, validationErr.FromGinError(err))
		return
	}

	cmd := authentication.LoginCommand{
		Email:      req.Email,
		Password:   req.Password,
		IP:         c.ClientIP(),
		UserAgent:  c.Request.UserAgent(),
		DeviceType: detectDeviceType(c.Request.UserAgent()),
	}

	resp, err := h.service.Login(c.Request.Context(), cmd)
	if err != nil {
		response.Error(c, err)
		return
	}

	response.Success(c, authentication.ToAuthResponse(resp))
}

// Logout 处理用户退出
// @Summary User logout
// @Tags Authentication
// @Accept json
// @Produce json
// @Security BearerAuth
// @Success 200 {object} map[string]interface{}
// @Failure 401 {object} response.ErrorResponse "Unauthorized"
// @Router /auth/logout [post]
func (h *AuthHandler) Logout(c *gin.Context) {
	userID := c.GetString("user_id")

	cmd := authentication.LogoutCommand{
		UserID: userID,
	}

	if err := h.service.Logout(c.Request.Context(), cmd); err != nil {
		response.Error(c, err)
		return
	}

	response.Success(c, gin.H{"message": "Logged out successfully"})
}

// RefreshToken 刷新 Access Token
// @Summary Refresh access token
// @Tags Authentication
// @Accept json
// @Produce json
// @Param request body RefreshTokenRequest true "Refresh token"
// @Success 200 {object} authentication.AuthResponse
// @Failure 400 {object} response.ErrorResponse
// @Failure 401 {object} response.ErrorResponse "Invalid or expired token"
// @Router /auth/refresh [post]
func (h *AuthHandler) RefreshToken(c *gin.Context) {
	var req RefreshTokenRequest
	if err := c.ShouldBindJSON(&req); err != nil {
		response.Error(c, validationErr.FromGinError(err))
		return
	}

	cmd := authentication.RefreshTokenCommand{
		RefreshToken: req.RefreshToken,
	}

	resp, err := h.service.RefreshToken(c.Request.Context(), cmd)
	if err != nil {
		response.Error(c, err)
		return
	}

	response.Success(c, authentication.ToAuthResponse(resp))
}

// GetUserByID 根据 ID 获取用户信息
// @Summary Get user by ID
// @Tags Users
// @Produce json
// @Security BearerAuth
// @Param id path string true "User ID"
// @Success 200 {object} authentication.UserResponse
// @Failure 401 {object} response.ErrorResponse "Unauthorized"
// @Failure 404 {object} response.ErrorResponse "User not found"
// @Router /users/{id} [get]
func (h *AuthHandler) GetUserByID(c *gin.Context) {
	userID := c.Param("id")
	if userID == "" {
		response.Error(c, errors.NewAppError(
			errors.ErrCodeSystemInvalidRequest,
			"用户 ID 不能为空",
			http.StatusBadRequest,
		))
		return
	}

	u, err := h.service.GetUserByID(c.Request.Context(), userID)
	if err != nil {
		response.Error(c, userErr.ErrNotFound)
		return
	}

	response.Success(c, authentication.ToUserResponse(u))
}

// GetCurrentUser 获取当前登录用户信息
// @Summary Get current user information
// @Tags Authentication
// @Produce json
// @Security BearerAuth
// @Success 200 {object} authentication.UserResponse
// @Failure 401 {object} response.ErrorResponse "Unauthorized"
// @Router /auth/me [get]
func (h *AuthHandler) GetCurrentUser(c *gin.Context) {
	userID := c.GetString("user_id")
	if userID == "" {
		response.Error(c, errors.NewAppError(
			errors.ErrCodeSystemUnauthorized,
			"缺少用户身份信息",
			http.StatusUnauthorized,
		))
		return
	}

	u, err := h.service.GetUserByID(c.Request.Context(), userID)
	if err != nil {
		response.Error(c, userErr.ErrNotFound)
		return
	}

	response.Success(c, authentication.ToUserResponse(u))
}

// maskIP 脱敏 IP 地址（支持 IPv4 和 IPv6）
func maskIP(ip string) string {
	if ip == "" {
		return ""
	}
	// IPv6 包含冒号
	if strings.Contains(ip, ":") {
		parts := strings.Split(ip, ":")
		if len(parts) >= 3 {
			return parts[0] + ":" + parts[1] + ":***"
		}
		return ip[:len(ip)/2] + "***"
	}
	// IPv4
	parts := strings.Split(ip, ".")
	if len(parts) == 4 {
		return parts[0] + "." + parts[1] + ".***"
	}
	return ip[:len(ip)/2] + "***"
}

// DeviceResponse 设备响应
type DeviceResponse struct {
	TokenID    string `json:"token_id"`
	DeviceType string `json:"device_type"`
	IP         string `json:"ip"`
	UserAgent  string `json:"user_agent"`
	CreatedAt  string `json:"created_at"`
	IsCurrent  bool   `json:"is_current"`
}

// DevicesResponse 设备列表响应
type DevicesResponse struct {
	Devices []DeviceResponse `json:"devices"`
}

// GetUserDevices 获取当前用户的所有登录设备
// @Tags Authentication
// @Produce json
// @Security BearerAuth
// @Success 200 {object} DevicesResponse
// @Failure 401 {object} response.ErrorResponse "Unauthorized"
// @Router /auth/devices [get]
func (h *AuthHandler) GetUserDevices(c *gin.Context) {
	userID := c.GetString("user_id")

	devices, err := h.tokenService.GetUserDevices(c.Request.Context(), userID)
	if err != nil {
		response.Error(c, errors.NewAppError(
			errors.ErrCodeSystemInternal,
			"获取设备列表失败",
			http.StatusInternalServerError,
		))
		return
	}

	// 通过 access_token → device_token 映射标识当前设备
	currentDeviceTokenID := ""
	authHeader := c.GetHeader("Authorization")
	if parts := strings.SplitN(authHeader, " ", 2); len(parts) == 2 && parts[0] == "Bearer" {
		if id, err := h.tokenService.GetCurrentDeviceTokenID(c.Request.Context(), parts[1]); err == nil {
			currentDeviceTokenID = id
		}
	}

	var deviceResponses []DeviceResponse
	for _, device := range devices {
		deviceResponses = append(deviceResponses, DeviceResponse{
			TokenID:    device.TokenID,
			DeviceType: device.DeviceType,
			IP:         maskIP(device.IP),
			UserAgent:  device.UserAgent,
			CreatedAt:  device.CreatedAt,
			IsCurrent:  device.TokenID == currentDeviceTokenID,
		})
	}

	response.Success(c, DevicesResponse{
		Devices: deviceResponses,
	})
}

// RevokeDevice 踢出指定设备
// @Summary Revoke a specific device
// @Tags Authentication
// @Produce json
// @Security BearerAuth
// @Param token path string true "Device token"
// @Success 200 {object} map[string]interface{}
// @Failure 400 {object} response.ErrorResponse
// @Failure 401 {object} response.ErrorResponse "Unauthorized"
// @Router /auth/devices/{token} [delete]
func (h *AuthHandler) RevokeDevice(c *gin.Context) {
	userID := c.GetString("user_id")
	token := c.Param("token")

	if token == "" {
		response.Error(c, errors.NewAppError(
			errors.ErrCodeSystemInvalidRequest,
			"设备令牌不能为空",
			http.StatusBadRequest,
		))
		return
	}

	deviceInfo, err := h.tokenService.ValidateRefreshTokenWithDevice(c.Request.Context(), token)
	if err != nil {
		response.Error(c, errors.NewAppError(
			errors.ErrCodeAuthInvalidToken,
			"无效或已过期的设备令牌",
			http.StatusUnauthorized,
		))
		return
	}

	if deviceInfo.UserID != userID {
		response.Error(c, errors.NewAppError(
			errors.ErrCodeSystemForbidden,
			"只能撤销自己的设备",
			http.StatusForbidden,
		))
		return
	}

	if err := h.tokenService.RevokeDeviceByToken(c.Request.Context(), token); err != nil {
		response.Error(c, errors.NewAppError(
			errors.ErrCodeSystemInternal,
			"撤销设备失败",
			http.StatusInternalServerError,
		))
		return
	}

	response.Success(c, gin.H{"message": "Device revoked successfully"})
}

// LogoutAllDevices 退出所有设备
// @Summary Logout from all devices
// @Tags Authentication
// @Produce json
// @Security BearerAuth
// @Success 200 {object} map[string]interface{}
// @Failure 401 {object} response.ErrorResponse "Unauthorized"
// @Router /auth/logout-all [post]
func (h *AuthHandler) LogoutAllDevices(c *gin.Context) {
	userID := c.GetString("user_id")

	if err := h.tokenService.RevokeAllDevices(c.Request.Context(), userID); err != nil {
		response.Error(c, errors.NewAppError(
			errors.ErrCodeSystemInternal,
			"退出所有设备失败",
			http.StatusInternalServerError,
		))
		return
	}

	response.Success(c, gin.H{"message": "Logged out from all devices successfully"})
}

// detectDeviceType 根据 User-Agent 检测设备类型
func detectDeviceType(userAgent string) string {
	if userAgent == "" {
		return "unknown"
	}

	if strings.Contains(userAgent, "iPad") || strings.Contains(userAgent, "Tablet") {
		return "tablet"
	}
	if strings.Contains(userAgent, "Mobile") || strings.Contains(userAgent, "Android") || strings.Contains(userAgent, "iPhone") {
		return "mobile"
	}

	return "desktop"
}